Change vmf

+++ cornac/models/vmf/vmf.py
@@ -19,7 +19,7 @@ except ImportError:
 def _load_or_randn(size, init_values, device):
     if init_values is None:
         tensor = np.random.normal(loc=0.0, scale=0.001, size=size)
-        tensor = torch.tensor(tensor, device=device, requires_grad=True)
+        tensor = torch.tensor(tensor, device=device, requires_grad=True, dtype=torch.float32)
     else:
         tensor = torch.tensor(init_values, requires_grad=True, device=device)
     return tensor
